requirements.txt
pyaudio
librosa
torch
pyyaml
numpy<2.0
ruff
rich
pytest